requirements.txt
bootstrap-admin==0.3.6
Django==1.8
Django-Accounts==0.1
django-admin-bootstrapped==2.5.0
django-allauth==0.24.1
django-appconf==1.0.1
django-audit==1.0b3
django-autoslug==1.7.2
django-bootstrap-calendar==0.1.0
django-bootstrap-form==3.2
django-bootstrap-toolkit==2.15.0
django-bootstrap3==5.3.1
django-chartit==0.1
django-crispy-forms==1.4.0
django-email-notification==1.3
django-forms-bootstrap==3.0.1
django-geojson==2.6.0
django-grappelli==2.7.1
django-leaflet==0.15.2
django-location-field==1.6.1
django-markdown==0.8.4
django-phonenumber-field==0.7.2
django-registration==1.0
django-registration-redux==1.1
django-suit==0.2.15
djangorestframework==3.0.5
easy-thumbnails==2.2
ecdsa==0.13
Fabric==1.10.1
icalendar==3.9.1
ipython==3.1.0
jsmin==2.1.6
Markdown==2.6
oauthlib==1.0.3
paramiko==1.15.2
phonenumbers==7.0.11
Pillow==2.7.0
pinax-theme-bootstrap==5.7.2
psycopg2==2.6
pycrypto==2.6.1
Pygments==2.0.2
Pylons==1.0.1
pymongo==3.1.1
python-dateutil==2.4.2
python-openid==2.2.5
pytz==2015.7
reportlab==3.2.0
requests==2.9.0
requests-oauthlib==0.6.0
simplejson==3.8.0
six==1.9.0
South==1.0.2
vectorformats==0.1
yolk==0.4.3
